[Examples] Merge python-simple and python-scipy
