+++ .ci/run-shellcheck
@@ -7,7 +7,7 @@ cd "$(git rev-parse --show-toplevel)"
 LC_ALL=C.UTF-8 shellcheck "$@" \
     Examples/bash/scripts/bash_test.sh \
     Examples/common_tools/benchmark-http.sh \
-    Examples/python-simple/run-tests.sh \
+    Examples/python/run-tests.sh \
     Runtime/pal_loader \
     Scripts/clean-check \
     Scripts/clean-check-prepare \
