--- Examples/python-scipy/python.manifest.template
@@ -1,94 +1,0 @@
-# Python3 manifest example
-#
-# This manifest was prepared and tested on Ubuntu 16.04.
-
-libos.entrypoint = "file:{{ entrypoint }}"
-
-# Path to the library OS
-loader.preload = "file:{{ graphene.libos }}"
-
-# Graphene log level
-loader.log_level = "{{ log_level }}"
-
-# Read application arguments directly from the command line. Don't use this on production!
-loader.insecure__use_cmdline_argv = 1
-
-# Environment variables for Python
-loader.env.LD_LIBRARY_PATH = "{{ python.stdlib }}/lib:/lib:{{ arch_libdir }}:/usr/lib:/usr/{{ arch_libdir }}"
-
-# Mounted FSes. The following "chroot" FSes mount a part of the host FS into the
-# guest. Other parts of the host FS will not be available in the guest.
-
-# Default glibc files, mounted from the Runtime directory in GRAPHENEDIR.
-fs.mount.lib.type = "chroot"
-fs.mount.lib.path = "/lib"
-fs.mount.lib.uri = "file:{{ graphene.runtimedir() }}"
-
-# Host-level libraries (e.g., /lib/x86_64-linux-gnu) required by the Python executable
-fs.mount.lib2.type = "chroot"
-fs.mount.lib2.path = "{{ arch_libdir }}"
-fs.mount.lib2.uri = "file:{{ arch_libdir }}"
-
-# Host-level directory (/usr) required by the Python executable
-fs.mount.usr.type = "chroot"
-fs.mount.usr.path = "/usr"
-fs.mount.usr.uri = "file:/usr"
-
-# Mount $PYTHONHOME
-fs.mount.pyhome.type = "chroot"
-fs.mount.pyhome.path = "{{ python.stdlib }}"
-fs.mount.pyhome.uri = "file:{{ python.stdlib }}"
-
-# Mount $PYTHONDISTHOME
-fs.mount.pydisthome.type = "chroot"
-fs.mount.pydisthome.path = "{{ python.distlib }}"
-fs.mount.pydisthome.uri = "file:{{ python.distlib }}"
-
-# Mount /tmp
-fs.mount.tmp.type = "chroot"
-fs.mount.tmp.path = "/tmp"
-fs.mount.tmp.uri = "file:/tmp"
-
-# Mount /etc
-fs.mount.etc.type = "chroot"
-fs.mount.etc.path = "/etc"
-fs.mount.etc.uri = "file:/etc"
-
-# Graphene general options
-
-# Graphene creates stacks of 256KB by default. It is not enough for SciPy/NumPy
-# packages, e.g., libopenblas dependency assumes more than 512KB-sized stacks.
-sys.stack.size = "2M"
-
-# SGX general options
-
-sgx.nonpie_binary = 1
-
-# Set the virtual memory size of the SGX enclave. For SGX v1, the enclave
-# size must be specified during signing. If Python needs more virtual memory
-# than the enclave size, Graphene will not be able to allocate it.
-sgx.enclave_size = "1G"
-
-# Set the maximum number of enclave threads. For SGX v1, the number of enclave
-# TCSes must be specified during signing, so the application cannot use more
-# threads than the number of TCSes. Note that Graphene also creates an internal
-# thread for handling inter-process communication (IPC), and potentially another
-# thread for asynchronous events. Therefore, the actual number of threads that
-# the application can create is (sgx.thread_num - 2).
-sgx.thread_num = 32
-
-# SGX trusted files
-
-sgx.trusted_files.python = "file:{{ entrypoint }}"
-
-sgx.trusted_files.runtime = "file:{{ graphene.runtimedir() }}/"
-sgx.trusted_files.arch_libdir = "file:{{ arch_libdir }}/"
-sgx.trusted_files.usr_libdir = "file:/usr/lib/" # can be removed after dropping support for
-                                                # Ubuntu 16.04
-sgx.trusted_files.python_dir = "file:{{ python.stdlib }}/"
-sgx.trusted_files.dist = "file:{{ python.distlib }}/"
-sgx.trusted_files.scripts = "file:scripts/"
-
-# SGX untrusted (allowed) files/directories
-sgx.allowed_files.tmp = "file:/tmp"
-sgx.allowed_files.etc = "file:/etc"
