--- Examples/python-simple/Makefile
@@ -1,63 +1,0 @@
-# Use one of the following commands to build the manifest for Python3:
-#
-# - make                Building for Linux
-# - make DEBUG=1        Building for Linux (with Graphene debug output)
-# - make SGX=1          Building for SGX
-# - make SGX=1 DEBUG=1  Building for SGX (with Graphene debug output)
-#
-# Use `make clean` to remove Graphene-generated files.
-
-include ../../Scripts/Makefile.configs
-
-# Relative path to Graphene root
-GRAPHENEDIR ?= ../..
-SGX_SIGNER_KEY ?= $(GRAPHENEDIR)/Pal/src/host/Linux-SGX/signer/enclave-key.pem
-
-ifeq ($(DEBUG),1)
-GRAPHENE_LOG_LEVEL = debug
-else
-GRAPHENE_LOG_LEVEL = error
-endif
-
-.PHONY: all
-all: python.manifest
-ifeq ($(SGX),1)
-all: python.manifest.sgx python.sig python.token
-endif
-
-python.manifest: python.manifest.template
-	graphene-manifest \
-		-Dlog_level=$(GRAPHENE_LOG_LEVEL) \
-		-Darch_libdir=$(ARCH_LIBDIR) \
-		-Dentrypoint=$(realpath $(shell sh -c "command -v python3")) \
-		$< >$@
-
-# Python manifests for SGX:
-#   Generating the SGX-specific manifest (python.manifest.sgx), the enclave signature,
-#   and the token for enclave initialization.
-
-python.manifest.sgx: python.manifest
-	graphene-sgx-sign \
-		--key $(SGX_SIGNER_KEY) \
-		--manifest $< \
-		--output $@
-
-python.sig: python.manifest.sgx
-
-python.token: python.sig
-	graphene-sgx-get-token --output $@ --sig $<
-
-.PHONY: check
-check: all
-	./run-tests.sh > TEST_STDOUT 2> TEST_STDERR
-	@grep -q "Success 1/3" TEST_STDOUT
-	@grep -q "Success 2/3" TEST_STDOUT
-	@grep -q "Success 3/3" TEST_STDOUT
-
-.PHONY: clean
-clean:
-	$(RM) *.manifest *.manifest.sgx *.token *.sig OUTPUT* *.PID TEST_STDOUT TEST_STDERR
-	$(RM) -r scripts/__pycache__
-
-.PHONY: distclean
-distclean: clean
